+++ plugins/tapd/api/connection.go
@@ -19,21 +19,16 @@ package api
 
 import (
 	"fmt"
-	"github.com/apache/incubator-devlake/config"
-	"github.com/apache/incubator-devlake/models/common"
-	"github.com/apache/incubator-devlake/plugins/tapd/models"
 	"net/http"
-	"strconv"
 	"time"
 
+	"github.com/apache/incubator-devlake/plugins/tapd/models"
+
 	"github.com/apache/incubator-devlake/plugins/helper"
-	"github.com/go-playground/validator/v10"
 	"github.com/mitchellh/mapstructure"
 
 	"github.com/apache/incubator-devlake/plugins/core"
 )
-
-var vld = validator.New()
 
 /*
 POST /plugins/tapd/test
@@ -78,131 +73,27 @@ func TestConnection(input *core.ApiResourceInput) (*core.ApiResourceOutput, erro
 	return nil, nil
 }
 
-func findConnectionByInputParam(input *core.ApiResourceInput) (*models.TapdConnection, error) {
-	connectionId := input.Params["connectionId"]
-	if connectionId == "" {
-		return nil, fmt.Errorf("missing connectionsid")
-	}
-	tapdConnectionId, err := strconv.ParseUint(connectionId, 10, 64)
-	if err != nil {
-		return nil, fmt.Errorf("invalid connectionId")
-	}
-
-	return getTapdConnectionById(tapdConnectionId)
-}
-
-func getTapdConnectionById(id uint64) (*models.TapdConnection, error) {
-	tapdConnection := &models.TapdConnection{}
-	err := db.First(tapdConnection, id).Error
-	if err != nil {
-		return nil, err
-	}
-
-	// decrypt
-	v := config.GetConfig()
-	encKey := v.GetString(core.EncodeKeyEnvStr)
-	tapdConnection.BasicAuthEncoded, err = core.Decrypt(encKey, tapdConnection.BasicAuthEncoded)
-	if err != nil {
-		return nil, err
-	}
-
-	return tapdConnection, nil
-}
-
-func mergeFieldsToTapdConnection(tapdConnection *models.TapdConnection, connections ...map[string]interface{}) error {
-	// decode
-	for _, connections := range connections {
-		err := mapstructure.Decode(connections, tapdConnection)
-		if err != nil {
-			return err
-		}
-	}
-
-	// validate
-	vld := validator.New()
-	err := vld.Struct(tapdConnection)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func refreshAndSaveTapdConnection(tapdConnection *models.TapdConnection, data map[string]interface{}) error {
-	var err error
-	// update fields from request body
-	err = mergeFieldsToTapdConnection(tapdConnection, data)
-	if err != nil {
-		return err
-	}
-
-	// encrypt
-	v := config.GetConfig()
-	encKey := v.GetString(core.EncodeKeyEnvStr)
-	if encKey == "" {
-		// Randomly generate a bunch of encryption keys and set them to config
-		encKey = core.RandomEncKey()
-		v.Set(core.EncodeKeyEnvStr, encKey)
-		err := config.WriteConfig(v)
-		if err != nil {
-			return err
-		}
-	}
-	tapdConnection.BasicAuthEncoded, err = core.Encrypt(encKey, tapdConnection.BasicAuthEncoded)
-	if err != nil {
-		return err
-	}
-
-	// transaction for nested operations
-	tx := db.Begin()
-	defer func() {
-		if err != nil {
-			tx.Rollback()
-		} else {
-			tx.Commit()
-		}
-	}()
-	if tapdConnection.RateLimit == 0 {
-		tapdConnection.RateLimit = 6480
-	}
-	if tapdConnection.ID > 0 {
-		err = tx.Save(tapdConnection).Error
-	} else {
-		err = tx.Create(tapdConnection).Error
-	}
-	if err != nil {
-		if common.IsDuplicateError(err) {
-			return fmt.Errorf("tapd connections with name %s already exists", tapdConnection.Name)
-		}
-		return err
-	}
-	tapdConnection.BasicAuthEncoded, err = core.Decrypt(encKey, tapdConnection.BasicAuthEncoded)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 /*
 POST /plugins/tapd/connections
 {
 	"name": "tapd data connections name",
 	"endpoint": "tapd api endpoint, i.e. https://merico.atlassian.net/rest",
-	"basicAuthEncoded": "generated by `echo -n <tapd login email>:<tapd token> | base64`",
+	"Token": "generated by `echo -n <tapd login email>:<tapd token> | base64`",
 	"rateLimit": 10800,
 }
 */
 func PostConnections(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
 	// create a new connections
-	tapdConnection := &models.TapdConnection{}
+	connection := &models.TapdConnection{}
 
 	// update from request and save to database
-	err := refreshAndSaveTapdConnection(tapdConnection, input.Body)
+	//err := refreshAndSaveTapdConnection(tapdConnection, input.Body)
+	err := connectionHelper.Create(connection, input)
 	if err != nil {
 		return nil, err
 	}
 
-	return &core.ApiResourceOutput{Body: tapdConnection, Status: http.StatusOK}, nil
+	return &core.ApiResourceOutput{Body: connection, Status: http.StatusOK}, nil
 }
 
 /*
@@ -210,53 +101,44 @@ PATCH /plugins/tapd/connections/:connectionId
 {
 	"name": "tapd data connections name",
 	"endpoint": "tapd api endpoint, i.e. https://merico.atlassian.net/rest",
-	"basicAuthEncoded": "generated by `echo -n <tapd login email>:<tapd token> | base64`",
+	"Token": "generated by `echo -n <tapd login email>:<tapd token> | base64`",
 	"rateLimit": 10800,
 }
 */
 func PatchConnection(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
-	// load from db
-	tapdConnection, err := findConnectionByInputParam(input)
+	connection := &models.TapdConnection{}
+	err := connectionHelper.Patch(connection, input)
 	if err != nil {
 		return nil, err
 	}
 
-	// update from request and save to database
-	err = refreshAndSaveTapdConnection(tapdConnection, input.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	return &core.ApiResourceOutput{Body: tapdConnection}, nil
+	return &core.ApiResourceOutput{Body: connection}, nil
 }
 
 /*
 DELETE /plugins/tapd/connections/:connectionId
 */
 func DeleteConnection(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
-	// load from db
-	tapdConnection, err := findConnectionByInputParam(input)
+	connection := &models.TapdConnection{}
+	err := connectionHelper.First(connection, input.Params)
 	if err != nil {
 		return nil, err
 	}
-	err = db.Delete(tapdConnection).Error
-	if err != nil {
-		return nil, err
-	}
-
-	return &core.ApiResourceOutput{Body: tapdConnection}, nil
+	err = connectionHelper.Delete(connection)
+	return &core.ApiResourceOutput{Body: connection}, err
 }
 
 /*
 GET /plugins/tapd/connections
 */
 func ListConnections(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
-	tapdConnections := make([]models.TapdConnection, 0)
-	err := db.Find(&tapdConnections).Error
+	var connections []models.TapdConnection
+	err := connectionHelper.List(&connections)
 	if err != nil {
 		return nil, err
 	}
-	return &core.ApiResourceOutput{Body: tapdConnections}, nil
+
+	return &core.ApiResourceOutput{Body: connections, Status: http.StatusOK}, nil
 }
 
 /*
@@ -266,18 +148,12 @@ GET /plugins/tapd/connections/:connectionId
 {
 	"name": "tapd data connections name",
 	"endpoint": "tapd api endpoint, i.e. https://merico.atlassian.net/rest",
-	"basicAuthEncoded": "generated by `echo -n <tapd login email>:<tapd token> | base64`",
+	"Token": "generated by `echo -n <tapd login email>:<tapd token> | base64`",
 	"rateLimit": 10800,
 }
 */
 func GetConnection(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
-	tapdConnection, err := findConnectionByInputParam(input)
-	if err != nil {
-		return nil, err
-	}
-
-	detail := &models.TapdConnectionDetail{
-		TapdConnection: *tapdConnection,
-	}
-	return &core.ApiResourceOutput{Body: detail}, nil
+	connection := &models.TapdConnection{}
+	err := connectionHelper.First(connection, input.Params)
+	return &core.ApiResourceOutput{Body: connection}, err
 }
